Fix: 타입 정의
