+++ components/letters/IconPicker.tsx
@@ -1,25 +1,16 @@
 import Image from "next/image";
-
-const icons = [
-  { id: 1, name: "bag", src: "/images/ic-bag.svg" },
-  { id: 2, name: "shoes", src: "/images/ic-shoes.svg" },
-  { id: 3, name: "dogtag", src: "/images/ic-byeoldol-face-blackhat.svg" },
-  { id: 4, name: "vest", src: "/images/ic-vest.svg" },
-  { id: 5, name: "face", src: "/images/ic-keyring.svg" },
-  { id: 6, name: "helmet", src: "/images/ic-helmet.svg" },
-  { id: 7, name: "gun", src: "/images/ic-gun.svg" },
-];
+import { ICONS, IconName } from "@/types/common/icons";
 
 export default function IconPicker({
   value,
   onChange,
 }: {
-  value: string;
-  onChange: (name: string) => void;
+  value: IconName;
+  onChange: (name: IconName) => void;
 }) {
   return (
     <div className="flex items-center gap-1">
-      {icons.map((icon) => (
+      {ICONS.map((icon) => (
         <button
           key={icon.id}
           type="button"
