+++ types/common/icons.ts
@@ -1,0 +1,24 @@
+export type IconName =
+  | "bag"
+  | "shoes"
+  | "dogtag"
+  | "vest"
+  | "face"
+  | "helmet"
+  | "gun";
+
+export type CabinetIcon = {
+  id: number;
+  name: IconName;
+  src: string;
+};
+
+export const ICONS: CabinetIcon[] = [
+  { id: 1, name: "bag", src: "/images/ic-bag.svg" },
+  { id: 2, name: "shoes", src: "/images/ic-shoes.svg" },
+  { id: 3, name: "dogtag", src: "/images/ic-byeoldol-face-blackhat.svg" },
+  { id: 4, name: "vest", src: "/images/ic-vest.svg" },
+  { id: 5, name: "face", src: "/images/ic-keyring.svg" },
+  { id: 6, name: "helmet", src: "/images/ic-helmet.svg" },
+  { id: 7, name: "gun", src: "/images/ic-gun.svg" },
+];
