+++ types/letters/letter.ts
@@ -1,0 +1,5 @@
+export type uploadedFileType = {
+  file: File;
+  url: string;
+  type: "image" | "video";
+};
